Mantis 0001075: o Fixed bug that made items on the popupmenu be disabled for aliens in classic games
The bug was in AvHPlayer::UpdateTechNodes(), where the techs would not be set to researchable once flagged as unresearchable, causing them all to be unresearchable from the start of the game, and disallowing lifeforms after gestating to them.

I also added a check so that the menu states won't be updated every frame for the client unless the menu is active.

git-svn-id: https://unknownworlds.svn.cloudforge.com/ns1@262 67975925-1194-0748-b3d5-c16f83f1a3a1